+++ Examples/Examples.sln
@@ -1,0 +1,42 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32825.248
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TimerExample", "Timer\TimerExample\TimerExample.csproj", "{E59FE9C8-6B68-4D16-A5D7-7D238366ED90}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Timer", "Timer", "{640410B1-7CB3-4482-9B51-959F46C5462C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TimerExample.Tests", "Timer\TimerExample.Tests\TimerExample.Tests.csproj", "{FA948873-1762-4B35-9D1B-0090B90F5CE9}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_", "_", "{641A4EEA-484D-4332-B410-45DE0E95A4C6}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.props = Directory.Build.props
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E59FE9C8-6B68-4D16-A5D7-7D238366ED90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E59FE9C8-6B68-4D16-A5D7-7D238366ED90}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E59FE9C8-6B68-4D16-A5D7-7D238366ED90}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E59FE9C8-6B68-4D16-A5D7-7D238366ED90}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FA948873-1762-4B35-9D1B-0090B90F5CE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FA948873-1762-4B35-9D1B-0090B90F5CE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FA948873-1762-4B35-9D1B-0090B90F5CE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FA948873-1762-4B35-9D1B-0090B90F5CE9}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(NestedProjects) = preSolution
+		{E59FE9C8-6B68-4D16-A5D7-7D238366ED90} = {640410B1-7CB3-4482-9B51-959F46C5462C}
+		{FA948873-1762-4B35-9D1B-0090B90F5CE9} = {640410B1-7CB3-4482-9B51-959F46C5462C}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D439CFB9-E2DD-4DF9-AA90-F321B50EAD16}
+	EndGlobalSection
+EndGlobal
