+++ Examples/Timer/TimerExample.Tests/TimerExample.Tests.csproj
@@ -1,0 +1,7 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <ItemGroup>
+    <ProjectReference Include="..\TimerExample\TimerExample.csproj" />
+  </ItemGroup>
+
+</Project>
